Ok, added the concept of an MVP, and it works only for one triangle. The reason is because I am reusing the same Heap location to store the current object transformation. This is a problem.
The solution is to create unique Constant Buffer data per renderable
object. This imposes a limitation of draw calls. Each Constant Buffer
can be as large as 64KB. And unique constant buffer views must be
aligned to 256Bytes. This means that if all of the data for a single
object fits in 256 Bytes, we can have 64KB/256B=256 unique renderable
objects per Constant Buffer.

Another example: If the per object data exceeds 256 Bytes but is less
than 512Bytes, then we can have 64KB/512B=128 unique renderable objects
per Constant Buffer.

So, the idea would be to create several constant buffers. Now, so far
we only store 64bytes per object which is less than 256, so we can store
256 matrices in each Constant buffer. And if we want to render thousands
of objects at once, we should create a constant buffer manager class
that owns all of the constant buffers.

Signed-off-by: Shuffle Master <shuffle_master@skyerocke.com>